+++ src/main/java/org/example/model/tipoCasilleros/Carcel.java
@@ -1,8 +1,8 @@
 package org.example.model.tipoCasilleros;
 public class Carcel extends Casillero{
-    private int fianza;
+    private double fianza;
     private int cantTurnos;
-    public Carcel(int ubicacion,int fianza,int cantTurnos) {
+    public Carcel(int ubicacion,double fianza,int cantTurnos) {
         super("Carcel",TipoCasillero.CARCEL,ubicacion);
         this.fianza = fianza;
         this.cantTurnos = cantTurnos;
